# mvcc/reader: reduce seek's count from default cf