TODO list:

- [ ] develop

- [ ] unit test

- [ ] preformace test

@BusyJay @hicqu PTAL