@jackysp 2018-01-18
250k rows, 51 mvcc versions for each row.

This pr:
```
MySQL [yusp]> select count(c) from sbtest1;
+----------+
| count(c) |
+----------+
|   250000 |
+----------+
1 row in set (0.62 sec)
```

master branch:
```
MySQL [yusp]> select count(c) from sbtest1;
+----------+
| count(c) |
+----------+
|   250000 |
+----------+
1 row in set (0.78 sec)
```