@AndreMouche 2018-01-19
## sysbench key with single version
### Environment

* 1 tikv/1 tidb/1 pd /on one machine
* 1 table with 5000000 records
* select from range with 256 thread

### Result

It seems no difference within the measurement uncertainties which meet expectation.

master

```
SQL statistics:
    queries performed:
        read:                            2822744
        write:                           0
        other:                           0
        total:                           2822744
    transactions:                        2822744 (9401.88 per sec.)
    queries:                             2822744 (9401.88 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          300.2287s
    total number of events:              2822744

Latency (ms):
         min:                                  2.04
         avg:                                 27.21
         max:                               1521.47
         95th percentile:                     45.79
         sum:                            76799167.62

Threads fairness:
    events (avg/stddev):           11026.3438/4971.88
    execution time (avg/stddev):   299.9967/0.04
```

current branch

```
SQL statistics:
    queries performed:
        read:                            2815973
        write:                           0
        other:                           0
        total:                           2815973
    transactions:                        2815973 (9379.03 per sec.)
    queries:                             2815973 (9379.03 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          300.2373s
    total number of events:              2815973

Latency (ms):
         min:                                  5.71
         avg:                                 27.27
         max:                               1479.53
         95th percentile:                     33.12
         sum:                            76799340.59

Threads fairness:
    events (avg/stddev):           10999.8945/1610.58
    execution time (avg/stddev):   299.9974/0.04
```