@MyonKeminta 2018-02-07
## Benchmark with different SEEK_BOUND

### Environment

* 1 tikv/1 tidb/1 pd /on one machine
* 1 table with 1000000 records
* select from range with 256 thread

### Results with single version

Branch     |SEEK_BOUND|QPS     
-------------|-------------------|----------
master     |30                    |27553  
master     |8                      |27623  
this pr      |30                    |27650  
this pr      |8                      |27417  

Seems this pr has almost no influence to that

### Results with multiple version

Branch     |SEEK_BOUND|QPS     
-------------|-------------------|----------
master     |30                    |5305     
master     |8                      |5663     
this pr      |30                    |6466     
this pr      |8                      |6530     

Even when set `SEEK_BOUND` to 8, this pr still improves the performance by 15%. However, changing `SEEK_BOUND` from 30 to 8 seems improves only a little.