@MyonKeminta 2018-02-07
Ah.. How if write_cf is filled with delete and rollback record? The keys in default_cf will be much less than those in write_cf...?