@MyonKeminta 2018-02-07
..Or, if most of the data is stored as short values, which means, the value will be stored with write records in write_cf but not default_cf. In these situations, keys in default_cf will be much sparser than those in write_cf. I think this PR may cause performance reduction in these cases.