@AndreMouche 2018-02-08
I do not think so, with short values, we needn't read from default cf. And In this PR, we only change the way to read default cf. So the performance shouldn't change for short values @MyonKeminta