@BusyJay 2018-02-08
I'm OK with closing it. But there are several points that are not accurate.

> This pr benefit when there are many versions for all rows...

Not necessary. As long as there are some keys that have many versions, it will benefit.

> ...and the query is range scan...

Range scan is very common.

> ...the benefit of this pr reduce.

According to the benchmark result from @MyonKeminta above, 15% is still a significant enhancement.